local-broker/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27,<0.40
httpx>=0.27
PyJWT>=2.8
websockets>=12.0
pydantic>=2.6
